% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_spatial.R
\name{build_gp_cholesky}
\alias{build_gp_cholesky}
\title{Build GP covariance Cholesky for given hyperparameters}
\usage{
build_gp_cholesky(dist_mat, range, sigma, nu)
}
\description{
Build GP covariance Cholesky for given hyperparameters
}
\keyword{internal}
