% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_laplace.R
\name{build_spatial_args}
\alias{build_spatial_args}
\title{Build spatial-specific arguments for C++ call}
\usage{
build_spatial_args(hp, spatial_info, n_spatial, has_shared)
}
\description{
Build spatial-specific arguments for C++ call
}
\keyword{internal}
