% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_spatial.R
\name{build_spde_precision}
\alias{build_spde_precision}
\title{Build SPDE precision matrix for given hyperparameters}
\usage{
build_spde_precision(spde_info, range, sigma)
}
\description{
Build SPDE precision matrix for given hyperparameters
}
\keyword{internal}
