% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{generate_quadrature}
\alias{generate_quadrature}
\title{Generate quadrature points for LGCP integral}
\usage{
generate_quadrature(loc_mat, sampling_spec, Z_samp)
}
\description{
Generate quadrature points for LGCP integral
}
\keyword{internal}
