% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{latent_correlated}
\alias{latent_correlated}
\title{Correlated Latent Structure}
\usage{
latent_correlated(rho_prior = NULL)
}
\arguments{
\item{rho_prior}{Prior on the correlation coefficient. Default is
uniform on (-1, 1).}
}
\value{
A \code{spjam_latent} object of type "correlated".
}
\description{
Specifies correlated (but not identical) spatial processes for ecology
and sampling. Estimates the correlation coefficient between the two
latent fields.
}
