% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_laplace.R
\name{make_param_names}
\alias{make_param_names}
\title{Generate parameter column names}
\usage{
make_param_names(model_data, has_shared)
}
\description{
Generate parameter column names
}
\keyword{internal}
