% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spjam_fit.R
\name{new_spjam_fit}
\alias{new_spjam_fit}
\title{Construct a spjam_fit object}
\usage{
new_spjam_fit(
  laplace_result,
  formula,
  family,
  spatial,
  shared,
  priors,
  data,
  locations,
  model_data,
  spatial_info,
  backend,
  seed
)
}
\description{
Construct a spjam_fit object
}
\keyword{internal}
