% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_half_cauchy}
\alias{prior_half_cauchy}
\title{Half-Cauchy Prior}
\usage{
prior_half_cauchy(location = 0, scale = 1)
}
\arguments{
\item{location}{Location parameter (typically 0).}

\item{scale}{Scale parameter.}
}
\value{
A \code{spjam_prior} object.
}
\description{
Half-Cauchy Prior
}
