% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_half_normal}
\alias{prior_half_normal}
\title{Half-Normal Prior}
\usage{
prior_half_normal(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Location (typically 0).}

\item{sd}{Scale parameter.}
}
\value{
A \code{spjam_prior} object.
}
\description{
Half-Normal Prior
}
