% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_normal}
\alias{prior_normal}
\title{Normal Prior}
\usage{
prior_normal(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Mean of the normal distribution.}

\item{sd}{Standard deviation.}
}
\value{
A \code{spjam_prior} object.
}
\description{
Normal Prior
}
