% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_pc_range}
\alias{prior_pc_range}
\title{PC Prior for Spatial Range}
\usage{
prior_pc_range(U, alpha = 0.5)
}
\arguments{
\item{U}{Upper bound for range.}

\item{alpha}{Tail probability.}
}
\value{
A \code{spjam_prior} object.
}
\description{
Penalized complexity prior for the practical range of a spatial process.
Specifies P(range < U) = alpha.
}
