% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_pc_sigma}
\alias{prior_pc_sigma}
\title{PC Prior for Marginal Standard Deviation}
\usage{
prior_pc_sigma(U, alpha = 0.5)
}
\arguments{
\item{U}{Upper bound.}

\item{alpha}{Tail probability.}
}
\value{
A \code{spjam_prior} object.
}
\description{
Penalized complexity prior for marginal standard deviation.
Specifies P(sigma > U) = alpha.
}
