% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling_lgcp}
\alias{sampling_lgcp}
\title{Log-Gaussian Cox Process Sampling Specification}
\usage{
sampling_lgcp(integration_points = 1000, boundary = NULL)
}
\arguments{
\item{integration_points}{Number of integration points for the intensity
integral. Default 1000.}

\item{boundary}{An sf polygon defining the study region boundary.}
}
\value{
A \code{spjam_sampling} object of type "lgcp".
}
\description{
Specifies a log-Gaussian Cox process (LGCP) model for the sampling
intensity. This is the default and most flexible option for point-referenced
observation data.
}
