% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling_point}
\alias{sampling_point}
\title{Point Process Sampling Specification}
\usage{
sampling_point(intensity_formula = ~1)
}
\arguments{
\item{intensity_formula}{Formula for the log-intensity.}
}
\value{
A \code{spjam_sampling} object of type "point".
}
\description{
Specifies an inhomogeneous Poisson point process model for sampling
locations. Simpler than LGCP but requires explicit intensity function.
}
