% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatr_fit.R
\name{scatr_fit}
\alias{scatr_fit}
\title{scatr_fit S3 class}
\description{
Output of \code{\link[=scatr]{scatr()}} containing posterior draws, model specifications,
and diagnostics from the Laplace approximation.
}
\keyword{internal}
