% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_car}
\alias{spatial_car}
\title{CAR Spatial Specification}
\usage{
spatial_car(adjacency = NULL, type = c("icar", "proper"))
}
\arguments{
\item{adjacency}{Adjacency matrix or neighborhood list.}

\item{type}{Type of CAR: "icar" (intrinsic) or "proper".}
}
\value{
A \code{spjam_spatial} object of type "car".
}
\description{
Specifies a conditional autoregressive (CAR) spatial effect for areal data.
}
