% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spjam-package.R
\docType{package}
\name{spjam-package}
\alias{spjam-package}
\title{spjam: Joint Spatial Modelling of Ecological Processes Under Preferential Sampling}
\description{
A Bayesian spatial modelling pipeline that separates ecological signal from sampling-driven scatter under preferential sampling. Implements joint hierarchical models where sampling locations and ecological processes are estimated simultaneously with shared latent spatial structure, producing sampling-adjusted predictions that reflect where ecological processes occur rather than where observers look.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gcol33/spjam}
  \item Report bugs at \url{https://github.com/gcol33/spjam/issues}
}

}
\author{
\strong{Maintainer}: Gilles Colling \email{gilles.colling05@gmail.com} [copyright holder]

}
\keyword{internal}
