% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spjam_fit.R
\name{spjam_fit}
\alias{spjam_fit}
\title{spjam_fit S3 class}
\description{
Output of \code{\link[=spjam]{spjam()}} containing posterior draws, model specifications,
and diagnostics from the Laplace approximation.
}
\keyword{internal}
