% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_laplace.R
\name{unpack_hyperparams}
\alias{unpack_hyperparams}
\title{Unpack hyperparameter vector into named list}
\usage{
unpack_hyperparams(theta, names, spatial_type, has_shared, family)
}
\description{
Unpack hyperparameter vector into named list
}
\keyword{internal}
