% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_scatr_inputs}
\alias{validate_scatr_inputs}
\title{Validate inputs to scatr()}
\usage{
validate_scatr_inputs(
  formula,
  data,
  locations,
  family,
  spatial,
  shared,
  sampling_spec
)
}
\description{
Validate inputs to scatr()
}
\keyword{internal}
